chore: wip

+++ src/main/java/com/ifa/b03/positif/entities/Address.java
@@ -18,8 +18,8 @@ public class Address {
     private Long id;
     
     private String address;
-    private Long longitude;
-    private Long latitude;
+    private double longitude;
+    private double latitude;
 
     public Address(String address) {
         this.address = address;
@@ -36,19 +36,19 @@ public class Address {
         this.address = address;
     }
 
-    public Long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public Long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
